fix(core): typo in mod.rs and rename payload_filter
