+++ src/core/mod.rs
@@ -2,7 +2,7 @@ pub mod disc;
 pub mod game_image;
 pub mod input_handler;
 pub mod input_registry;
-pub mod payloa_filter;
+pub mod payload_filter;
 pub mod platform;
 pub mod reader;
 pub mod reader_factory;
